move all apis to data module
