+++ data/src/main/kotlin/com/vultisig/wallet/data/api/ApiModule.kt
@@ -12,6 +12,35 @@ internal interface ApiModule {
 
     @Binds
     @Singleton
+    fun bindOneInchApi(
+        impl: OneInchApiImpl
+    ): OneInchApi
+
+    @Binds
+    @Singleton
+    fun bindVultiSignerApi(
+        impl: VultiSignerApiImpl
+    ): VultiSignerApi
+
+    @Binds
+    @Singleton
+    fun bindEvmApi(
+        impl: EvmApiFactoryImp,
+    ): EvmApiFactory
+
+    @Binds
+    @Singleton
+    fun bindPolkadotApi(
+        impl: PolkadotApiImp
+    ): PolkadotApi
+
+    @Binds
+    fun bindSessionApi(
+        impl: SessionApiImpl
+    ): SessionApi
+
+    @Binds
+    @Singleton
     fun bindCoinGeckoApi(
         impl: CoinGeckoApiImpl,
     ): CoinGeckoApi
