+++ data/src/main/kotlin/com/vultisig/wallet/data/api/CoinGeckoApi.kt
@@ -1,21 +1,17 @@
 package com.vultisig.wallet.data.api
 
 import com.vultisig.wallet.data.models.Chain
-import com.vultisig.wallet.data.utils.BigDecimalSerializer
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
 import io.ktor.client.request.header
 import io.ktor.client.request.parameter
-import io.ktor.client.statement.bodyAsText
-import kotlinx.serialization.builtins.MapSerializer
-import kotlinx.serialization.builtins.serializer
 import java.math.BigDecimal
 import javax.inject.Inject
 
-internal typealias CurrencyToPrice = Map<String, BigDecimal>
+typealias CurrencyToPrice = Map<String, BigDecimal>
 
-internal interface CoinGeckoApi {
+interface CoinGeckoApi {
 
     suspend fun getCryptoPrices(
         priceProviderIds: List<String>,
