+++ data/src/main/kotlin/com/vultisig/wallet/data/api/errors/SwapException.kt
@@ -1,6 +1,6 @@
 package com.vultisig.wallet.data.api.errors
 
-internal sealed class SwapException(message: String): Exception(message){
+sealed class SwapException(message: String): Exception(message){
     class SwapIsNotSupported(message: String): SwapException(message)
     class AmountCannotBeZero(message: String): SwapException(message)
     class SameAssets(message: String): SwapException(message)
