+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/BlowfishResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class BlowfishResponse(
+data class BlowfishResponse(
     @SerialName("warnings")
     val warnings: List<BlowfishWarning>?,
     @SerialName("aggregated")
@@ -12,13 +12,13 @@ internal data class BlowfishResponse(
 )
 
 @Serializable
-internal data class BlowfishAggregated(
+data class BlowfishAggregated(
     @SerialName("warnings")
     val warnings: List<BlowfishWarning>?,
 )
 
 @Serializable
-internal data class BlowfishWarning(
+data class BlowfishWarning(
     @SerialName("message")
     val message: String?,
 )
